refactore folder structure + update npm packages + webpack build (without type:module) + mocha tests (with type:module)
