+++ auth/config.js
