+++ package.json
@@ -19,31 +19,31 @@
     "url": "https://github.com/david-88/todo/issues"
   },
   "homepage": "https://github.com/david-88/todo#readme",
-  "type": "module",
+
   "devDependencies": {
-    "@babel/core": "^7.12.10",
-    "@babel/plugin-transform-runtime": "^7.12.10",
-    "@babel/preset-env": "^7.12.11",
+    "@babel/core": "^7.15.0",
+    "@babel/plugin-transform-runtime": "^7.15.0",
+    "@babel/preset-env": "^7.15.0",
     "babel-cli": "^6.26.0",
     "babel-core": "^6.26.3",
     "babel-loader": "^8.2.2",
     "babel-preset-env": "^1.7.0",
-    "css-loader": "^5.0.1",
-    "eslint": "^7.19.0",
-    "eslint-config-standard": "^16.0.2",
-    "eslint-plugin-import": "^2.22.1",
+    "css-loader": "^5.2.7",
+    "eslint": "^7.32.0",
+    "eslint-config-standard": "^16.0.3",
+    "eslint-plugin-import": "^2.24.1",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^4.2.1",
-    "mocha": "^9.0.1",
+    "mocha": "^9.1.0",
     "uuid": "^8.3.2",
-    "webpack": "^5.4.0",
-    "webpack-bundle-analyzer": "^4.3.0",
-    "webpack-cli": "^4.2.0",
+    "webpack": "^5.51.1",
+    "webpack-bundle-analyzer": "^4.4.2",
+    "webpack-cli": "^4.8.0",
     "webpack-dev-server": "^3.11.1"
   },
   "dependencies": {
-    "firebase": "^8.2.1",
-    "firebaseui": "^4.7.1",
+    "firebase": "^8.10.0",
+    "firebaseui": "^4.8.1",
     "style-loader": "^2.0.0"
   }
 }
